+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Fixed
-- Fix design issues on telemarketing popover
+- Fix design issues on telemarketing popover and bar
 
 ## [2.0.0] - 2018-11-08
 ### Changed