+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.0.1] - 2018-11-27
+### Fixed
+- Fix design issues on telemarketing popover and bar
+
 ## [2.0.0] - 2018-11-08
 ### Changed
 - Migrate app to typescript.